`native:tail` select device target and log file
